[EventLoop] Associate only one listener per each stream event type.
Allowing multiple listeners makes code more complex and it is beyond
the scope of the whole event loop abstraction anyway.

Such a requirement could be easily satisfied on a higher level if a
class wrapping the event loop needs to notify more than one listener
about changes in the status of a stream resource.