Prototype of new loop that allows enabling and disabling listeners
This design has a major advantage: The definition of the listener and the
enabling are completely separate. Quite often you want to disable reads
as a pushback mechanism until your own buffer has space again. Similarly,
you only want to listen for writes when there is something to write.

Being able to enable and disable listeners without having to re-add the
actual listener callable is a lot more comfortable than having to juggle
the callbacks and re-add them all the time.