Skip signal tests w/o PCNTL and skip timer tests on inaccurate platforms
The signal constants are only available when ext-pcntl is loaded. If it
is not loaded, simply skip all tests that reference these constants.

Before running a timer test and asserting its time interval, run a
simple test to check the accuracy on the current platform. This is known
to work on most common platforms, but will now avoid false negative test
results on platforms that are known to be slow (Travis CI and others).