Check PCNTL functions for signal support instead of PCNTL extension
This helps avoiding problems when the PCNTL extension is loaded, but
(some of) its functions are disabled via PHPs `disable_function`
configuration. This is particularly common for the PCNTL functions in
CGI environments, e.g. they are disabled by default on Debian- /
Ubuntu-based distributions. Previously, even when not attaching any
signal listeners, the loop would print a warning on each loop tick:

> PHP Warning:  pcntl_signal_dispatch() has been disabled for security
reasons in {file} on {line}